refactor(editor): group keyboard memory plugin files
